+++ guns-admin/src/main/java/com/stylefeng/guns/common/persistence/model/Relation.java
@@ -21,53 +21,53 @@ public class Relation extends Model<Relation> {
     /**
      * 主键
      */
-	@TableId(value="id", type= IdType.AUTO)
-	private Integer id;
+    @TableId(value = "id", type = IdType.AUTO)
+    private Integer id;
     /**
      * 菜单id
      */
-	private Integer menuid;
+    private Long menuid;
     /**
      * 角色id
      */
-	private Integer roleid;
+    private Integer roleid;
 
 
-	public Integer getId() {
-		return id;
-	}
+    public Integer getId() {
+        return id;
+    }
 
-	public void setId(Integer id) {
-		this.id = id;
-	}
+    public void setId(Integer id) {
+        this.id = id;
+    }
 
-	public Integer getMenuid() {
-		return menuid;
-	}
+    public Long getMenuid() {
+        return menuid;
+    }
 
-	public void setMenuid(Integer menuid) {
-		this.menuid = menuid;
-	}
+    public void setMenuid(Long menuid) {
+        this.menuid = menuid;
+    }
 
-	public Integer getRoleid() {
-		return roleid;
-	}
+    public Integer getRoleid() {
+        return roleid;
+    }
 
-	public void setRoleid(Integer roleid) {
-		this.roleid = roleid;
-	}
+    public void setRoleid(Integer roleid) {
+        this.roleid = roleid;
+    }
 
-	@Override
-	protected Serializable pkVal() {
-		return this.id;
-	}
+    @Override
+    protected Serializable pkVal() {
+        return this.id;
+    }
 
-	@Override
-	public String toString() {
-		return "Relation{" +
-			"id=" + id +
-			", menuid=" + menuid +
-			", roleid=" + roleid +
-			"}";
-	}
+    @Override
+    public String toString() {
+        return "Relation{" +
+                "id=" + id +
+                ", menuid=" + menuid +
+                ", roleid=" + roleid +
+                "}";
+    }
 }
